+++ python/tests/test_surface_integral.py
@@ -95,20 +95,11 @@ def test_surface_integrals(get_assemblers):  # noqa: F811
     fem.petsc.set_bc(b, bcs)
 
     solver.setOperators(A)
-    uh = b.copy()
-    uh.set(0)
-    solver.solve(b, uh)
-    uh.ghostUpdate(addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD)
+    uh = fem.Function(mpc.function_space)
+    uh.x.set(0)
+    solver.solve(b, uh.vector)
+    uh.x.scatter_forward()
     mpc.backsubstitution(uh)
-
-    # Write solution to file
-    # u_h = dolfinx.Function(mpc.function_space)
-    # u_h.vector.setArray(uh.array)
-    # u_h.name = "u_mpc"
-    # outfile = dolfinx.io.XDMFFile(MPI.COMM_WORLD, "output/uh.xdmf", "w")
-    # outfile.write_mesh(mesh)
-    # outfile.write_function(u_h)
-    # outfile.close()
 
     # Solve the MPC problem using a global transformation matrix
     # and numpy solvers to get reference values
@@ -130,7 +121,7 @@ def test_surface_integrals(get_assemblers):  # noqa: F811
         A_csr = dolfinx_mpc.utils.gather_PETScMatrix(A_org, root=root)
         K = dolfinx_mpc.utils.gather_transformation_matrix(mpc, root=root)
         L_np = dolfinx_mpc.utils.gather_PETScVector(L_org, root=root)
-        u_mpc = dolfinx_mpc.utils.gather_PETScVector(uh, root=root)
+        u_mpc = dolfinx_mpc.utils.gather_PETScVector(uh.vector, root=root)
 
         if MPI.COMM_WORLD.rank == root:
             KTAK = K.T * A_csr * K